+++ cli/cmd/silero_vad.onnx
